git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@23973 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/RequestStarterGroup.java
@@ -202,9 +202,9 @@ public class RequestStarterGroup {
 		MyRequestThrottle throttle = getThrottle(isSSK, isInsert);
 		sb.append(TimeUtil.formatTime((long)throttle.getRTT(), 2, true));
 		sb.append(" delay=");
-		sb.append(TimeUtil.formatTime((long)throttle.getDelay(), 2, true));
+		sb.append(TimeUtil.formatTime(throttle.getDelay(), 2, true));
 		sb.append(" bw=");
-		sb.append((long)throttle.getRate());
+		sb.append(throttle.getRate());
 		sb.append("B/sec");
 		return sb.toString();
 	}
